requirements/base.txt
beautifulsoup4==4.9.3
django-filter==2.4.0
djangorestframework==3.12.2
django-cacheops==5.1
django-cors-headers==3.6.0
django-chunkator==2.0.0
django-extensions==3.0.9
lxml==4.6.2
mysqlclient==2.0.1
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-forked==1.0.2
pytest-mock==1.10.4
pytest-ordering==0.6
pytest-xdist==1.29.0
Django==2.2.17
requests==2.25.0
furl==2.1.0
sentry-sdk==0.19.5
slackweb==1.0.5
